+++ src/services/proxy-server/nginx-manager.ts
@@ -11,6 +11,7 @@ import { NginxServerConf } from './conf/nginx-server-conf';
 import { SSLManager } from './ssl-manager';
 import { DomainRepository } from '../../repositories/domain-repository';
 import ServerUtils from '../../utils/server';
+import Iptables, { Chain, Target } from '../../utils/iptables';
 
 export class NginxManager {
   static async addDomainServer(
@@ -161,6 +162,21 @@ export class NginxManager {
   ): Promise<void> {
     await this.removeLocation(service.domain, service.pathLocation);
 
+    const rule = {
+      chain: Chain.OUTPUT,
+      outInterface: service.node.wgInterface,
+      destination: service.node.address,
+      protocol: 'tcp',
+      dport: `${service.backendPort}`,
+      target: Target.REJECT,
+      args: {
+        '--reject-with': 'tcp-reset',
+      },
+    };
+
+    Iptables.addRule(rule);
+    Iptables.deleteRule(rule);
+
     if (restart) {
       await this.reloadServer();
     }
@@ -260,6 +276,21 @@ export class NginxManager {
     await FileManager.removeFile(
       `/etc/nginx/stream.d/${service.identifier}.conf`,
     );
+
+    const rule = {
+      chain: Chain.OUTPUT,
+      outInterface: service.node.wgInterface,
+      destination: service.node.address,
+      protocol: 'tcp',
+      dport: `${service.backendPort}`,
+      target: Target.REJECT,
+      args: {
+        '--reject-with': 'tcp-reset',
+      },
+    };
+
+    Iptables.addRule(rule);
+    Iptables.deleteRule(rule);
 
     if (restart) {
       await this.reloadServer();
